Remove some unused variables; add some headers that were missing; fix some returns that were void or not void inappropriately; change a couple of variables to size_t
git-svn-id: http://svn.net-core.org/repos/t-engine4@628 51575b47-30f0-44d4-a5cc-537603b46e54

+++ src/music.c
@@ -26,6 +26,8 @@
 #include "music.h"
 #include "script.h"
 #include "tSDL.h"
+#include "physfs.h"
+#include "physfsrwops.h"
 
 bool sound_active = TRUE;
 
@@ -63,7 +65,7 @@ static int music_play(lua_State *L)
 	int loop = lua_isnumber(L, 2) ? lua_tonumber(L, 2) : 1;
 	int fadein = lua_isnumber(L, 3) ? lua_tonumber(L, 3) : 0;
 
-	printf("play music %x %d %d\n", *m, loop, fadein);
+	printf("play music %x %d %d\n", (unsigned int)(*m), loop, fadein);
 	lua_pushboolean(L, (Mix_FadeInMusic(*m, loop, fadein) == -1) ? FALSE : TRUE);
 	return 1;
 }
